video convert

+++ src/main/java/com/uduck/mv/controller/UserController.java
@@ -3,8 +3,12 @@ package com.uduck.mv.controller;
 import com.uduck.mv.entity.Video;
 import com.uduck.mv.entity.dto.VideoDTO;
 import com.uduck.mv.entity.form.DataPage;
+import com.uduck.mv.security.MvUserDetails;
 import com.uduck.mv.service.IVideoService;
+import com.uduck.mv.util.ResponseResult;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -33,15 +37,38 @@ public class UserController {
     }
 
     @PutMapping("/user/video")
-    public String updateVideo(Video video){
-        videoService.addVideo(video);
-        return "success";
+    @ResponseBody
+    public ResponseResult updateVideo(Video video){
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (principal != null && principal instanceof MvUserDetails){
+            int operatorId = ((MvUserDetails) principal).getId();
+            VideoDTO videoDto = videoService.getVideoDtoById(video.getId());
+            if (operatorId == videoDto.getUser().getId()){
+                videoService.addVideo(video);
+                return ResponseResult.success();
+            } else {
+                return ResponseResult.fail(ResponseResult.Status.FORBIDDEN);
+            }
+        } else {
+            return ResponseResult.fail(ResponseResult.Status.FORBIDDEN);
+        }
     }
 
     @DeleteMapping("/user/video/{id}")
     @ResponseBody
-    public String deleteVideo(@PathVariable("id") Integer id){
-        videoService.deleteVideoById(id);
-        return "success";
+    public ResponseResult deleteVideo(@PathVariable("id") Integer id){
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (principal != null && principal instanceof MvUserDetails){
+            int operatorId = ((MvUserDetails) principal).getId();
+            VideoDTO videoDto = videoService.getVideoDtoById(id);
+            if (operatorId == videoDto.getUser().getId()){
+                videoService.deleteVideoById(id);
+                return ResponseResult.success();
+            } else {
+                return ResponseResult.fail(ResponseResult.Status.FORBIDDEN);
+            }
+        } else {
+            return ResponseResult.fail(ResponseResult.Status.FORBIDDEN);
+        }
     }
 }
